requirements.txt
aiogram>=3.3.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pytz>=2023.3
